requirements.txt
attrs==22.1.0
brotlipy==0.7.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi @ file:///C:/Windows/Temp/abs_6808y9x40v/croots/recipe/cffi_1659598653989/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama==0.4.5
contourpy==1.0.5
cryptography @ file:///C:/ci/cryptography_1652083563162/work
cycler==0.11.0
fonttools==4.37.4
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184308/work
mkl-service==2.4.0
numpy @ file:///C:/Windows/Temp/abs_2a1e1vbeag/croots/recipe/numpy_and_numpy_base_1659432712056/work
packaging==21.3
Pillow==9.2.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.8.2
pytorch-metric-learning==1.6.2
requests @ file:///C:/ci/requests_1657735342357/work
scikit-learn==1.1.2
scipy==1.9.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions @ file:///C:/Windows/TEMP/abs_dd2d0moa85/croots/recipe/typing_extensions_1659638831135/work
urllib3 @ file:///C:/Windows/TEMP/abs_65ynz4fdmi/croots/recipe/urllib3_1659110473919/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
